Moved 'hot-keys' and 'environment' to common
