+++ core/common/assets/js/utils/environment.js
