+++ core/common/assets/js/utils/hot-keys.js
